ml_model/requirements.txt
flask
keras==2.8.0
numpy==1.22.3
opencv-python==4.5.5.64
scikit-learn==1.0.2
tensorflow==2.8.0
aiohttp
Pillow
imageio